pyproject.toml
[project]
name = "crew-media-ops"
version = "0.1.0"
description = "Multi-agent system for social media operations using CrewAI"
authors = [
    {name = "Yiran Li", email = "1808848492@qq.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
license = {text = "MIT"}

dependencies = [
    # CrewAI core (without tools to avoid lancedb on Windows)
    "crewai>=0.80.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.11.0",
    "tenacity>=9.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.20",
    "orjson>=3.10.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    # Additional CrewAI dependencies
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    # Note: chromadb requires lancedb which doesn't support Windows
    # Use duckdb+ducktrie or sqlite for vector storage instead
    "duckdb>=1.0.0",
    # Security dependencies
    "cryptography>=44.0.0",
    "python-jose[cryptography]>=3.3.0",
    # Redis for caching (optional but recommended)
    "redis>=5.0.0",
    "alembic>=1.18.4",
    "aiofiles>=24.1.0",
    "markdown>=3.10.2",
    "playwright>=1.58.0",
    "python-socketio[asyncio]>=5.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
browser = [
    "playwright>=1.48.0",
    "websocket-client>=1.8.0",  # For CDP websocket fallback
]
overseas = [
    # Overseas platform SDKs (recommended for production)
    "praw>=7.8.0",           # Reddit
    "tweepy>=4.14.0",        # Twitter/X
    "instagrapi>=2.1.0",     # Instagram
    "facebook-sdk>=3.1.0",   # Facebook Graph API
    "threadspipepy>=0.1.0",  # Threads API
]

[project.urls]
Homepage = "https://github.com/yiranli/crew-media-ops"
Repository = "https://github.com/yiranli/crew-media-ops"
Documentation = "https://github.com/yiranli/crew-media-ops/wiki"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Windows platform constraint for uv
[tool.uv]
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
]

[tool.hatch.build.targets.wheel]
packages = ["src/crew"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "W191",   # tabs
    "RUF012", # mutable class attributes
    "RUF001", # ambiguous unicode chars in strings (Chinese project)
    "RUF002", # ambiguous unicode chars in docstrings (Chinese project)
    "RUF003", # ambiguous unicode chars in comments (Chinese project)
    "ARG002", # unused method arguments (common in interface implementations)
    "ARG003", # unused class method arguments (validators need info param)
]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["crew"]
force-sort-within-sections = true
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG", "S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "-v",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "crewai.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\b",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\\.pyi?$'
exclude = '''
/(
    \\.git
    | \\.mypy_cache
    | \\.ruff_cache
    | \\.venv
    | \\.eggs
    | \\.tox
    | build
    | dist
)/
'''
